+++ server/test.py
@@ -1,6 +1,7 @@
 from models.config import engine
 from models.config import session
 from app.add_db_LendInfo import AddLendInfoData
+from app.get_db import GetLendData
 from datetime import datetime ,date, timedelta
 from models.lend_info import Lend_info
 
@@ -17,7 +18,7 @@ deadline_date_string = deadline_data.strftime('%Y/%m/%d %H:%M:%S')
 
 print(now_date_string,deadline_date_string)
 
-#AddLendInfoData(user_id_data,borrower_id_data,book_id_data,deadline_data)
+AddLendInfoData(user_id_data,borrower_id_data,book_id_data,deadline_data)
 
 
 now_date = datetime.now()
@@ -27,16 +28,5 @@ a =  str(tomorrow)
 print(type(a))
 print(str(now_date))
 print(str(tomorrow))
-
-def GetLendData(id):
-    print(type(id))
-    select = (id)
-    print("select",select)
-    users = session.query(Lend_info).filter(Lend_info.id==id).all()
-    session.commit()
-    if users ==[]:
-        return None
-    for row in users:
-        return [row.id,row.borrower_id,row.own_book_id,row.created_at,row.returned_at,row.deadline,row.is_valid]
 
 print(GetLendData(1))
